Iterate over Ifinfomsg attributes once when listing names
The `get_attr_payload_as` method iterates over attributes until a
matching type is found. Calling this repeatedly for each desired
attribute type causes results in an O(n^2) runtime.

These messages are very small, so avoiding `get_attr_payload_as` may be
a bit of a pre-mature optimization, but the code before the neli 0.5.3
upgrade avoided this. Making this change to keep existing behavior on
the main branch.